Intel(R) SGX DCAP 1.22 Release
Upgraded Intel DCAP Quote Verification Enclave to integrate OpenSSL/SgxSSL 3.0.14.
Removed Intel DCAP PCCS from repository.
Added Ubuntu* 24.04 LTS 64-bit Server support.
Fixed bugs.

Note that PCCS is not available from this release. Please follow DCAP installation guide to use
`PCCSAdminTool` to retrieve the attestation collaterals or use old version PCCS.

+++ tools/SGXPlatformRegistration/network/src/MPSynchronicSender.cpp
@@ -59,7 +59,7 @@ static size_t responseCallback(void *b, size_t size, size_t nmemb, void *useptr)
     uint16_t chunkSize = (uint16_t)(size*nmemb);
     struct Buffer *mem = (struct Buffer*)useptr;
 
-    if (mem->pos + chunkSize < chunkSize) {
+    if ((uint16_t)(mem->pos + chunkSize) < chunkSize) {
         network_log_message_aux(mpNetworkLoglevel, MP_REG_LOG_LEVEL_ERROR, "Integer overflow happens during processing reponse.\n");
         return 0;
     }
